src/config.py
"""Central configuration for the FLM Agent pipeline."""

from pathlib import Path

from dotenv import load_dotenv
from pydantic import Field
from pydantic_settings import BaseSettings

load_dotenv()

# Paths
ROOT_DIR = Path(__file__).resolve().parent.parent
DATA_DIR = ROOT_DIR / "data"
RAW_DATA_DIR = DATA_DIR / "raw"
PROCESSED_DIR = DATA_DIR / "processed"
EVALUATION_DIR = ROOT_DIR / "evaluation"


class Settings(BaseSettings):
    """Application settings loaded from environment variables."""

    # OpenAI
    openai_api_key: str = Field(default="")
    embedding_model: str = "text-embedding-3-small"
    embedding_dimensions: int = 1536
    chat_model: str = "gpt-4o-mini"

    # Pinecone
    pinecone_api_key: str = Field(default="")
    pinecone_index_name: str = "flm-agent"

    # LangSmith
    langsmith_api_key: str = Field(default="")

    # Chunking
    chunk_size: int = 500
    chunk_overlap: int = 50

    # Retrieval
    top_k: int = 5

    model_config = {"env_file": ".env", "extra": "ignore"}


settings = Settings()
