pyproject.toml
[project]
name = "flm-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "jupyterlab>=4.5.5",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.1.3",
    "langsmith>=0.7.22",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "pandas>=1.4.0",
    "pinecone-client>=6.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "seaborn>=0.13.2",
    "streamlit>=1.45",
    "tiktoken>=0.12.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
